Refactor player statistics calculations and move utility functions to a separate module
